Refactoring.
1. Rename `PagingInstance` to `Paging`
2. Merge `paging_instance.*` into `paging.*`
3. Fix some docs.
