+++ include/kernel/kernel/kernel.hpp
@@ -30,11 +30,10 @@ class Kernel final {
     // The kernel singleton.
     static std::unique_ptr<Kernel> kernel_;
 
-    [[nodiscard]] static memory::paging::PagingInstance
-    initialize_kernel_paging();
+    [[nodiscard]] static memory::paging::Paging initialize_kernel_paging();
 
     constexpr static size_t MAX_NUMBER_OF_DISKS = 4;
 
     std::unique_ptr<drivers::storage::Disk> disks_[MAX_NUMBER_OF_DISKS];
-    memory::paging::PagingInstance kernel_paging_;
+    memory::paging::Paging kernel_paging_;
 };
